Add files via upload
I think I finally got this to a working state! Later, I want to implement an automatic data flushing feature once the CSV file reaches beyond a certain amount. This would essentially be deleting the first row of the CSV file when writing new data once it reaches a row cap (say, 10000 or something) to limit file size. I haven't yet figured a way to do so without using the Pandas library. Seeing as this is built on Python-standard lists and databases, idk if I want to introduce all that bloat and rewrite the entire script just to have a one-line solution to data deletion. I'm instead going to see what options the built-in Python CSV module has.