Add files via upload
Included with this modified Database file is the graph of my first ascent test with the first janky variation of what will become the gradient ascent algorithm.

It didn't work as planned, which I expected. Actually, all previous attempts to do a gradient ascent with manual data led to my beam entirely decoupling from the fiber cable, and this did not happen. It just got stuck. That's a good thing.

You'll note that an extra button was added. All this button does is call to the ascent program. I was hoping to keep the database entirely separate from the auto-alignment, but USB limitations mean that I can't have two python programs requesting data from the LabJack at the same time. If I want to monitor AND adjust, it has to be on the same program. To preserve the database's functionality even without a Picomotor in place, I'm just going to have the ascent option be inaccessible without a Picomotor attached to your computer. For now this is just a "pass" function if the computer cant find a Picomotor in place, but eventually I'll make the button grey itself out if no Picomotor is present (assuming I choose to keep it as a button rather than implementing the ascent functionality in some other way).

It feels like there are already too many buttons.